Vavr replaced with Arrow

+++ src/main/kotlin/com/github/grambbledook/example/spark/dto/Account.kt
@@ -1,3 +1,5 @@
 package com.github.grambbledook.example.spark.dto
 
-data class Account(val id: Long, val amount: Double, val owner: String)
+import java.math.BigDecimal
+
+data class Account(val id: Long, val amount: BigDecimal, val owner: String)
